componentizing, start dead code removal
