--- web/src/actions/index.js
@@ -1,5 +1,0 @@
-export const INCREMENT_COUNT = 'INCREMENT_COUNT';
-export const DECREMENT_COUNT = 'DECREMENT_COUNT';
-
-export const increment = () => ({ type: INCREMENT_COUNT });
-export const decrement = () => ({ type: DECREMENT_COUNT });
