+++ web/src/components/v2/CertifyAndSubmit.js
@@ -1,0 +1,16 @@
+import React from 'react';
+
+import Section from './Section';
+import SectionTitle from './SectionTitle';
+import Collapsible from '../Collapsible';
+
+const CertifyAndSubmit = () => (
+  <Section>
+    <SectionTitle>Certify and Submit</SectionTitle>
+    <Collapsible title="Certify and Submit">
+      <div>...</div>
+    </Collapsible>
+  </Section>
+);
+
+export default CertifyAndSubmit;
