+++ web/src/components/v2/HelpBox.js
@@ -1,0 +1,22 @@
+import PropTypes from 'prop-types';
+import React from 'react';
+
+import Icon, { faHelpSolid } from '../Icons';
+
+const HelpBox = ({ children }) => (
+  <div className="my2 p2 h5 sm-col-7 lg-col-6 bg-teal-light border border-teal border-width-2 rounded relative">
+    <div
+      className="inline-block absolute line-height-1 bg-white circle"
+      style={{ top: '-.5em', left: '-.5em' }}
+    >
+      <Icon icon={faHelpSolid} className="teal" />
+    </div>
+    {children}
+  </div>
+);
+
+HelpBox.propTypes = {
+  children: PropTypes.node.isRequired
+};
+
+export default HelpBox;
