+++ web/src/components/v2/TopNav.js
@@ -1,0 +1,32 @@
+import PropTypes from 'prop-types';
+import React from 'react';
+
+import Icon, { faHelp, faSignOut } from '../Icons';
+
+const TopNav = ({ place }) => (
+  <header className="clearfix py1 bg-white">
+    <div className="sm-col">
+      <a href="#!" className="btn caps">
+        2018 {place.name} HITECH APD
+      </a>
+    </div>
+    <div className="sm-col-right h5">
+      <div>
+        <button type="button" className="btn h5 regular">
+          <span className="mr-tiny">Help</span>
+          <Icon icon={faHelp} />
+        </button>
+        <button type="button" className="btn h5 regular">
+          <span className="mr-tiny">Log out</span>
+          <Icon icon={faSignOut} />
+        </button>
+      </div>
+    </div>
+  </header>
+);
+
+TopNav.propTypes = {
+  place: PropTypes.object.isRequired
+};
+
+export default TopNav;
